# Conflicts:
#	bin/test/Main.class
#	bin/tries/TriesHybrides.class
#	src/test/Main.java

J'avais ajouté la listeMots(), j'espère ne rien casser.

+++ src/tries/Trie.java
@@ -8,6 +8,6 @@ public interface Trie {
 	public String[] listeMots();
 	public int comptageNil();
 	public int hauteur();
-	public int profondeurMoyenne();
+	public double profondeurMoyenne();
 	public int prefixe(String mot);
 }
